+++ packages/arb-token-bridge-ui/src/token-bridge-sdk/index.ts
@@ -1,19 +1,19 @@
-export * from './hooks/useArbTokenBridge'
-export * from './hooks/arbTokenBridge.types'
+export * from '../hooks/useArbTokenBridge'
+export * from '../hooks/arbTokenBridge.types'
 
-export { useBalance } from './hooks/useBalance'
+export { useBalance } from '../hooks/useBalance'
 
 export type {
   Transaction,
   TxnStatus,
   NewTransaction,
   TxnType
-} from './hooks/useTransactions'
+} from '../hooks/useTransactions'
 
-export { txnTypeToLayer } from './hooks/useTransactions'
+export { txnTypeToLayer } from '../hooks/useTransactions'
 export type {
   L1ToL2MessageData,
   L2ToL1MessageData
-} from './hooks/useTransactions'
+} from '../hooks/useTransactions'
 
-export { useGasPrice } from './hooks/useGasPrice'
+export { useGasPrice } from '../hooks/useGasPrice'
